refactor all web app stuff into web app folder
