+++ webapp/index.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <head>
-  <link type="text/css" rel="stylesheet" href="./webapp/puzz.css">
+  <link type="text/css" rel="stylesheet" href="./puzz.css">
 </head>
 <body>
   <!-- weird hack to prevent flash of unstyled content on page load -->
